Improved Input Handling, NMS Plugin improvements
* Redesign of input management. Per Frame update mouse and keyboard states are passed to the engine and the UI controller. The window is still uses keyboard and mouse event handlers to update the current state that is pushed on frame update. Implemented custom mouse and keyboard structs to enable abstraction.
* Queued engine input states are removed. One state is polled whenever required.
* Scrolling issues in ImGui are also fixed.
* Added cleaner index length calculation mechanism during NMS mesh imports (hopefully there are no exceptions).

+++ NibbleCore/Platform/OpenGL/Graphics/OpenGLAPI.cs
@@ -584,7 +584,7 @@ namespace NbCore.Platform.Graphics.OpenGL
                         throw new NotImplementedException();
                 }
 
-                GL.VertexAttribPointer(buf.semantic, buf.count, buftype, buf.normalize, buf.stride, buf.offset);
+                GL.VertexAttribPointer(buf.semantic, buf.count, buftype, buf.normalize, (int) buf.stride, buf.offset);
                 GL.EnableVertexAttribArray(buf.semantic);
             }
